MIPS: ingenic: Add support for new JZ4770 based boards
Add support for the YLM (known as the Anbernic in the western
world) RG-280V, RG-280M, RG-300X, RG-350, RG-350M, and the Wolsen
PocketGo2 (aka. PlayGo) Rev. 1 & 2.

The YLM RG-280V is almost a carbon copy of the GCW-Zero, with a
different LCD panel (320x480 non-square pixels in diamond pattern),
L2/R2 and Vol +/- buttons added, and with the analog stick, ACT8600
regulator, RDA5807 radio, MXC6225 accelerometer, ITE6610 HDMI chips
removed.

The YLM RG-280M v1.1 is a RG-280V with a clickable analog stick. The
v1.0 additionally has a ITE66121 chip.

The Wolsen PocketGo2 Rev. 1 & 2 are similar to the RG-280V, with a
proper 320x240 IPS screen this time.

The RG-350 is a RG-280M v1.0 with two clickable analog sticks, and
a proper 320x240 IPS screen.

The RG-350M is a RG-350 with a 640x480 IPS screen.

The RG-300X is a RG-350 without joysticks.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>