drm/ingenic: Lookup for the proper clock where to register notifier
The code in place was making sure that when the rate of the pixel
clock's parent changed, the driver was notified and the pixel clock
could be recomputed.

This works fine unless the pixel clock has the CLK_SET_RATE_PARENT flag.
In that case, calling clk_set_rate() on the pixel clock will actually
change the rate of the parent clock, which triggers the notifier, and
causes the pixel clock's rate to be recomputed, effectively creating an
endless loop.

To fix this, instead of registering a notifier on the parent, we go up
the hierarchy tree and register a notifier on the parent of the first
clock that does not have the CLK_SET_RATE_PARENT set.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>