MIPS: ingenic: Add support for the Wolsen LDK
The Wolsen LDK are small handhelds with a 320x480 4:3 TFT LCD screen
(non-square pixels).

Two versions exist, one with a vertical layout (buttons below the
screen) and one with a horizontal layout (buttons on the sides of the
screen). Apart from that difference, the vertical version uses the
JZ4760B SoC, while the horizontal version uses the JZ4760 SoC.

Both are carbon copies of the RS-97 board, except for the different SoC
(for the vertical LDK) and a different panel used.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>