drm: ingenic: Add dma clock
While the LCD controller uses its own built-in DMA engine to transfer
the video frames instead of the SoC's DMA controller, it does need the
DMA controller's clock to be enabled.

From the point of view of the driver the dma clock is optional, in order
to keep compatibility with older Device Tree which did not provide it,
and worked because the driver for the DMA controller kept the clock
enabled.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>