iio: core: Add support for IIO_AVAIL_LIST_WITH_TYPE
The IIO_AVAIL_LIST_WITH_TYPE specifies that the array that corresponds
to the available values is composed by cells of 3 integers, the first
two representing the value itself (similar to what you would have with
IIO_AVAIL_LIST), and the third integer representing the encoding type of
the value.

This can be used for instance when a driver's .read_avail() callback
returns values which cannot be represented with an unique encoding type.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>