MIPS: ingenic: gcw0: Force peripheral mode on USB OTG
The GCW Zero's ID detect mechanism has a tendency to toast itself,
causing it to misdetect the USB mode and dumping current on the wires
when it's not supposed to, causing further damage elsewhere.
To support devices with the toasted ID detection, as well as prevent
potential hardware damage on yet unaffected devices, force USB mode
to peripheral only.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>